refactor(finances): split monolithic controller and implement reactive signals
- Separate FinancesController into Budget, Category, Expense, and Installment controllers.
- Implement auto_create_budget_for_wedding signal handler.
- Refactor BudgetService to support headless budget setup for signals.
