+++ backend/apps/finances/apps.py
@@ -4,3 +4,6 @@ from django.apps import AppConfig
 class FinancesConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "apps.finances"
+
+    def ready(self):
+        import apps.finances.signals  # noqa
